First version of importer and exporter that kind-of works. Not recommended for major work, yet. You should back up your existing Hurrican data before trying this. For advanced users only. No warranty. etc.
HurricanObjects.py is a shared database with object information
extracted from the original editor.

TiledToHurrican.py attempts to convert a level built with Tiled
into a level that works with Hurrican. I recommend looking at
the documentation for the original editor for information about
attributes.

HurricanToTiled.py attempts to convert a Hurrican level into a
map that can be edited in Tiled.

T2002ToTiled.py is an old test script to convert a T2002 level
into Tiled format. I'm most likely not to work on this further.

attribute_icons.png and attribute_icons.tsx is an external tileset
for Tiled to mark up the background for Hurrican.

objecttypes.xml can be imported into Tiled so that object types
are available in the drop-down box.